+++ HISTORY.md
@@ -6,6 +6,7 @@
 * Make per-request expiration thread-safe for both `CachedSession.request()` and `CachedSession.send()`
 * Handle some additional corner cases when normalizing request data
 * Some micro-optimizations for request matching
+* Fix license metadata as shown on PyPI
 
 ## 0.8.1 (2021-09-15)
 * Redact `ingored_parameters` from `CachedResponse.url` (if used for credentials or other sensitive info)