+++ HISTORY.md
@@ -7,6 +7,8 @@
 * Make per-request expiration thread-safe for both `CachedSession.request()` and `CachedSession.send()`
 * Handle some additional corner cases when normalizing request data
 * Some micro-optimizations for request matching
+* Fix issue with cache headers not being used correctly if `cache_control=True` is used with an `expire_after` value
+* Support immediate expiration + revalidation for `Cache-Control: max-age=0` and `Expires: 0`
 * Fix license metadata as shown on PyPI
 
 ## 0.8.1 (2021-09-15)