+++ HISTORY.md
@@ -7,6 +7,7 @@
 * Use `Cache-Control` **request** headers by default
 * Add support for `Cache-Control: immutable`
 * Add support for immediate expiration + revalidation with `Cache-Control: max-age=0` and `Expires: 0`
+* Reset expiration for cached response when a `304 Not Modified` response is received
 
 **Backends:**
 * Filesystem and SQLite backends: Add better error message if parent path exists but isn't a directory